Splited StrategyTab component into Sandbox and Live strategy tabs
