+++ src/components/StrategyEditor/tabs/StrategyLiveTab.js
