+++ src/components/StrategyEditor/tabs/StrategySandboxTabSandbox.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const StrategySandboxTab = () => {
+  return <div></div>
+}
